Add a script that imports keepassx2 csv databases
To make it easier to put the configuration in the click context (needed
in write_db()), now read_config() does not rely on the the context
itself to get the configuration path, but rather gets it as an argument.
It also returns the whole configuration and not just the user's, with
the user-defined options overwritting the default ones.