Allow matching the window title with more keywords
Add support for an optional `keywords` field in entries. When using
autotype, in addition to the entry name, the contents of this field are
also searched for in the window's title. `keywords` should contain a
string or a list of strings, but anything that is not a string is
converted to one.