Improve `passata generate --clipboard` behavior
If an existing password is overwritten, the user probably wants to
change his password and needs to fill a form which requires the old
password once and the new password twice. So the old password is put to
clipboard available for one paste, a prompt to press any key is given,
and then the new password is put to clipboard available for two pastes.